requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.10.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.5
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
scrapy==2.11.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.0
transformers==4.35.0
pydantic[email]==2.11.7

